Use uint64_t for TxnId

+++ common/types.h
@@ -12,7 +12,7 @@ namespace slog {
 
 using Key = std::string;
 using Value = std::string;
-using TxnId = uint32_t;
+using TxnId = uint64_t;
 using BatchId = uint32_t;
 using SlotId = uint32_t;
